requirements.txt
questionary
pyfiglet
colorama
click
pyyaml
exifread
